feat: implement Promo Code & Discount Engine feature
- Fix type mismatch: convert double to BigDecimal in RiderServiceImpl and ScheduledRideServiceImpl
- Add DistanceServiceOSRMImpl.calculateDistanceAndDuration() for ETA calculation
- Create PromoCode and PromoCodeUsage entities
- Create PromoCodeRepository and PromoCodeUsageRepository
- Create PromoCodeDto, PromoCodeApplyDto, PromoCodeResultDto DTOs
- Create PromoCodeService interface and PromoCodeServiceImpl implementation
- Supports FLAT and PERCENTAGE discount types
- Validates promo codes: active, date range, usage limit, minimum fare, per-user usage
- Create PromoCodeController with admin CRUD and rider apply endpoint

+++ ubercloneapp/src/test/java/com/github/k2ocabhinav/ubercloneapp/services/impl/DriverServiceImplTest.java
@@ -32,6 +32,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -102,7 +103,7 @@ class DriverServiceImplTest {
                 .driver(testDriver)
                 .rideStatus(RideStatus.CONFIRMED)
                 .otp("1234")
-                .fare(100.0)
+                .fare(BigDecimal.valueOf(100.0))
                 .paymentMethod(PaymentMethod.CASH)
                 .build();
     }
